+++ dart/lib/src/transport/http_transport.dart
@@ -6,7 +6,6 @@ import 'package:http/http.dart';
 import '../noop_client.dart';
 import '../protocol.dart';
 import '../sentry_options.dart';
-import '../utils.dart';
 import 'noop_encode.dart' if (dart.library.io) 'encode.dart';
 import 'transport.dart';
 
@@ -30,7 +29,8 @@ class HttpTransport implements Transport {
 
   HttpTransport._(this._options)
       : _dsn = Dsn.parse(_options.dsn!),
-        _headers = _buildHeaders(_options.sdk.identifier) {
+        _headers = _buildHeaders(
+            _options.platformChecker.isWeb, _options.sdk.identifier) {
     _credentialBuilder = _CredentialBuilder(
       _dsn,
       _options.sdk.identifier,
@@ -138,7 +138,7 @@ class _CredentialBuilder {
   }
 }
 
-Map<String, String> _buildHeaders(String sdkIdentifier) {
+Map<String, String> _buildHeaders(bool isWeb, String sdkIdentifier) {
   final headers = {'Content-Type': 'application/json'};
   // NOTE(lejard_h) overriding user agent on VM and Flutter not sure why
   // for web it use browser user agent