+++ flutter/lib/src/web_platform_checker.dart
@@ -1,5 +1,0 @@
-/// always false for flutter web
-bool isIOS() => false;
-
-/// always false for flutter web
-bool isAndroid() => false;