+++ flutter/test/mocks.dart
@@ -1,7 +1,45 @@
 import 'package:mockito/annotations.dart';
 import 'package:sentry/sentry.dart';
+import 'package:sentry/src/platform/platform.dart';
 
 const fakeDsn = 'https://abc@def.ingest.sentry.io/1234567';
 
 @GenerateMocks([Hub, Transport])
 void main() {}
+
+class MockPlatform implements Platform {
+  MockPlatform({
+    String? os,
+    String? osVersion,
+    String? hostname,
+  })  : operatingSystem = os ?? '',
+        operatingSystemVersion = osVersion ?? '',
+        localHostname = hostname ?? '';
+
+  @override
+  String operatingSystem;
+
+  @override
+  String operatingSystemVersion;
+
+  @override
+  String localHostname;
+
+  @override
+  bool get isLinux => (operatingSystem == 'linux');
+
+  @override
+  bool get isMacOS => (operatingSystem == 'macos');
+
+  @override
+  bool get isWindows => (operatingSystem == 'windows');
+
+  @override
+  bool get isAndroid => (operatingSystem == 'android');
+
+  @override
+  bool get isIOS => (operatingSystem == 'ios');
+
+  @override
+  bool get isFuchsia => (operatingSystem == 'fuchsia');
+}