+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Feat: Support for Linux (#402)
 * Feat: Support for Windows (#407)
 * Fix: Mark `Sentry.currentHub` as deprecated (#406)
+* Fix: Set console logger as default logger in debug mode (#413)
 * Fix: Use name from pubspec.yaml for release if package id is not available (#411)
 * Feat: `SentryHttpClient` tracks the duration which a request takes and logs failed requests (#414)
 * Bump: sentry-cocoa to v7.0.0 (#424)